chore: PHP 8.4 modernization — #[Override], array_* helpers, property hooks
Adopt PHP 8.4 idioms across src/:

- Add #[Override] to every interface/parent implementation that lacked it
  (118 methods across 82 files; full coverage now). Validated by reflection,
  PHPStan, and the suite — PHP fatals on a wrong attribute.
- Replace foreach search/existence/universal loops over plain arrays with
  array_find / array_any / array_all where it reads more clearly.
- Convert pure boolean accessors to get-hook virtual properties, matching the
  existing DocComment::$empty convention: ResponseNode (isSuccess/isError/
  isRedirect/isDefault), RequestBodyNode (isMultipart/isJson), LintOptions
  (isScoped), FixResult (hasChanges), OverrideMatcher (hasOverrides), and
  ResourceTarget (isAmbiguous). Hooked properties may not be readonly, so the
  four value objects that were `final readonly class` move the readonly modifier
  from the class onto each property — immutability is unchanged. SpecDefinition
  keeps isDefault() as a method (no callers, nothing to gain).

Also fixes a latent bug surfaced mid-pass: a dropped (array) cast in
ValidationRulesToSchema::normaliseIndexedPaths broke colliding paths carrying
a Closure rule. No change to generated documents or lint output.

+++ src/Support/Extraction/FakerExampleSynthesiser.php
@@ -72,7 +72,7 @@ final readonly class FakerExampleSynthesiser
         }
 
         if ($this->seed !== null) {
-            $this->faker->seed($this->seed ^ (int) crc32($fieldName));
+            $this->faker->seed($this->seed ^ crc32($fieldName));
         }
 
         if ($descriptor->enum !== null && $descriptor->enum !== []) {
